Refactor app initialization process to enhance clarity and functionality
- Introduced a new `runInitialization` function to encapsulate the full initialization process, allowing for better management within React components.
- Updated the `App` component to utilize the new initialization function, improving the handling of app state during loading.
- Simplified the `InitSpinner` component by removing unnecessary state checks and focusing on the current initialization step.
- Enhanced logging throughout the initialization process for improved debugging and monitoring.
