+++ apps/pwa/src/main.tsx
@@ -1,10 +1,6 @@
-import { initServices } from "@/app/services/init-services.ts";
-import { useAppStore } from "@/shared/stores/app-store.tsx";
-import { initStores } from "@/shared/stores/init-stores.ts";
 import { useInitializationStore, loadInitializationLog } from "@/shared/stores/initialization-store.tsx";
 import { PwaRegister } from "@/app/providers/pwa-register";
 import { AuthProvider } from "@/app/providers/auth-provider";
-import { runUnifiedMigrations, hasPendingMigrations } from "@/db/migrations";
 import { logger } from "@/shared/lib/logger.ts";
 import { Buffer } from "buffer";
 import { ConvexProvider, ConvexReactClient } from "convex/react";
@@ -50,31 +46,31 @@ function ConvexWrapper({ children }: { children: React.ReactNode }) {
 }
 
 /**
- * Initialize the application
+ * Bootstrap the application
  *
- * Key design: We render App ONCE and let it handle initialization UI internally.
- * This prevents HMR conflicts that occur when root.render() is called multiple times
- * with different component trees.
+ * Key design: We render App ONCE and let it handle initialization internally.
+ * The actual initialization (PGlite, services, stores) is triggered by App component
+ * after React has mounted and stabilized. This ensures:
+ * 1. React is fully mounted before DB operations
+ * 2. iOS Safari has time to stabilize storage context after redirects
+ * 3. Initialization runs within React lifecycle, not in module scope
  */
-async function initializeApp() {
-  // Track initialization time
-  const startTime = performance.now();
+function bootstrap() {
+  logger.debug("🚀 Bootstrapping application...");
 
   // Initialize steps in the store BEFORE rendering
-  const { initializeSteps, startStep, completeStep, warnStep, failStep, saveLog, setShowProgressScreen } =
-    useInitializationStore.getState();
+  const { initializeSteps, setShowProgressScreen } = useInitializationStore.getState();
 
   // Check previous initialization log
   const previousLog = loadInitializationLog();
   const isFirstInstall = !previousLog;
-  const previousInitSuccessful = previousLog && !previousLog.hasError;
 
   // First install: show detailed progress screen immediately
   if (isFirstInstall) {
     setShowProgressScreen(true);
   }
 
-  // Define all initialization steps
+  // Define all initialization steps (these will be run by App component)
   initializeSteps([
     { id: "database-engine", label: "Initialize database engine" },
     { id: "database-init", label: "Initialize database" },
@@ -100,10 +96,11 @@ async function initializeApp() {
     { id: "extensions", label: "Initialize extensions" },
   ]);
 
-  // Create root and render App AFTER store is initialized
+  // Create root and render App
   const root = createRoot(document.getElementById("root")!);
 
-  // Render app once - App component handles showing InitializationScreen vs actual app
+  // Render app - App component handles showing InitializationScreen vs actual app
+  // and triggers initialization after React has mounted
   root.render(
     <StrictMode>
       <AuthProvider>
@@ -115,123 +112,7 @@ async function initializeApp() {
     </StrictMode>,
   );
 
-  // OAuth callback special handling:
-  // Skip ALL initialization on /auth/callback to prevent PGlite hang on iOS Chrome/Safari.
-  // The AuthCallback component will handle token exchange and redirect to a clean URL.
-  const isAuthCallback = window.location.pathname === "/auth/callback";
-  if (isAuthCallback) {
-    logger.debug("🔐 OAuth callback detected - skipping initialization, will init after redirect");
-    // Mark app as initialized so AuthCallback component can render and process tokens
-    useAppStore.getState().setIsAppInitialized(true);
-    return;
-  }
-
-
-  // Progress callback for all initialization functions
-  const onProgress = (
-    stepId: string,
-    status: "start" | "success" | "warning" | "error",
-    error?: string,
-  ) => {
-    if (status === "start") {
-      startStep(stepId);
-    } else if (status === "success") {
-      completeStep(stepId);
-    } else if (status === "warning") {
-      warnStep(stepId, error || "Unknown warning");
-    } else if (status === "error") {
-      failStep(stepId, error || "Unknown error");
-    }
-    // App component subscribes to store changes via Zustand - no re-render needed
-  };
-
-  // Variable to track if migrations were needed (for logging)
-  let needsMigration = false;
-
-  try {
-    // Wait for render to stabilize before initializing DB
-    // This helps iOS Safari stabilize storage context after redirects
-    await new Promise((resolve) => requestAnimationFrame(() => setTimeout(resolve, 50)));
-
-    // Request persistent storage to reduce eviction risk on iOS
-    try {
-      const persisted = await navigator.storage?.persist?.();
-      logger.debug(`Storage persistence: ${persisted ? "granted" : "denied or unavailable"}`);
-    } catch {
-      logger.debug("Storage persistence API not available");
-    }
-
-    // Step 1: Initialize database engine (PGlite)
-    // This can take ~2 seconds on first load (PGlite initialization polling)
-    onProgress("database-engine", "start");
-
-    const dbCheckStart = performance.now();
-    needsMigration = await hasPendingMigrations();
-    const dbCheckDuration = performance.now() - dbCheckStart;
-
-    logger.debug(`⏱️ DB initialization took ${Math.round(dbCheckDuration)}ms`);
-    onProgress("database-engine", "success");
-
-    // App update with new migrations: upgrade to progress screen
-    if (needsMigration && !isFirstInstall) {
-      setShowProgressScreen(true);
-    }
-
-    logger.debug(`🔍 Pending migrations: ${needsMigration}`);
-
-    // Step 2: Migrate database (only if there are pending migrations)
-    if (needsMigration) {
-      logger.debug("🔨 Running database migrations...");
-      await runUnifiedMigrations(onProgress);
-    } else {
-      logger.debug("⏭️ No pending migrations, skipping migration steps");
-      // Mark migration steps as success immediately
-      onProgress("database-init", "start");
-      onProgress("database-init", "success");
-      onProgress("migration-schema", "start");
-      onProgress("migration-schema", "success");
-      onProgress("check-migrations", "start");
-      onProgress("check-migrations", "success");
-      onProgress("run-migrations", "start");
-      onProgress("run-migrations", "success");
-    }
-
-    // Step 3: Init services
-    logger.debug("🔧 Initializing services...");
-    await initServices(onProgress);
-
-    // Step 4: Init stores
-    logger.debug("📦 Initializing stores...");
-    await initStores(onProgress);
-
-    // Calculate initialization time
-    const initTime = performance.now() - startTime;
-    logger.debug(`✅ Initialization completed in ${Math.round(initTime)}ms`);
-
-    // Save initialization log on first install or when migrations were executed
-    if (isFirstInstall || needsMigration) {
-      saveLog(Math.round(initTime));
-    }
-
-    // Navigate to home on first install (before router initializes)
-    // This ensures users start from "/" after initial setup, regardless of URL
-    if (isFirstInstall && window.location.pathname !== "/") {
-      window.history.replaceState(null, "", "/");
-    }
-
-    // Mark app as initialized - App component will switch from InitializationScreen to actual app
-    useAppStore.getState().setIsAppInitialized(true);
-  } catch (error) {
-    logger.error("Failed to initialize app:", error);
-
-    // On failure, persist the current step states as a log
-    if (isFirstInstall || needsMigration) {
-      const initTime = performance.now() - startTime;
-      saveLog(Math.round(initTime));
-    }
-
-    // Show progress screen with error state
-    setShowProgressScreen(true);
-  }
+  logger.debug("✅ Bootstrap complete, App component will handle initialization");
 }
-initializeApp();
+
+bootstrap();
